Move profile types to new file
